+++ helpers_test.go
@@ -28,9 +28,17 @@ func TestAttributesFromDSN(t *testing.T) {
 		expected []attribute.KeyValue
 	}{
 		{
+			dsn: "mysql://root:otel_password@example.com/db",
+			expected: []attribute.KeyValue{
+				semconv.ServerAddress("example.com"),
+				semconv.DBNamespace("db"),
+			},
+		},
+		{
 			dsn: "mysql://root:otel_password@tcp(example.com)/db?parseTime=true",
 			expected: []attribute.KeyValue{
 				semconv.ServerAddress("example.com"),
+				semconv.DBNamespace("db"),
 			},
 		},
 		{
@@ -38,6 +46,7 @@ func TestAttributesFromDSN(t *testing.T) {
 			expected: []attribute.KeyValue{
 				semconv.ServerAddress("example.com"),
 				semconv.ServerPort(3307),
+				semconv.DBNamespace("db"),
 			},
 		},
 		{
@@ -45,18 +54,21 @@ func TestAttributesFromDSN(t *testing.T) {
 			expected: []attribute.KeyValue{
 				semconv.ServerAddress("2001:db8:1234:5678:9abc:def0:0001"),
 				semconv.ServerPort(3307),
+				semconv.DBNamespace("db"),
 			},
 		},
 		{
 			dsn: "mysql://root:otel_password@tcp(2001:db8:1234:5678:9abc:def0:0001)/db?parseTime=true",
 			expected: []attribute.KeyValue{
 				semconv.ServerAddress("2001:db8:1234:5678:9abc:def0:0001"),
+				semconv.DBNamespace("db"),
 			},
 		},
 		{
 			dsn: "root:secret@tcp(mysql)/db?parseTime=true",
 			expected: []attribute.KeyValue{
 				semconv.ServerAddress("mysql"),
+				semconv.DBNamespace("db"),
 			},
 		},
 		{
@@ -64,16 +76,20 @@ func TestAttributesFromDSN(t *testing.T) {
 			expected: []attribute.KeyValue{
 				semconv.ServerAddress("mysql"),
 				semconv.ServerPort(3307),
+				semconv.DBNamespace("db"),
 			},
 		},
 		{
-			dsn:      "root:secret@/db?parseTime=true",
-			expected: nil,
+			dsn: "root:secret@/db?parseTime=true",
+			expected: []attribute.KeyValue{
+				semconv.DBNamespace("db"),
+			},
 		},
 		{
 			dsn: "example.com/db?parseTime=true",
 			expected: []attribute.KeyValue{
 				semconv.ServerAddress("example.com"),
+				semconv.DBNamespace("db"),
 			},
 		},
 		{
@@ -81,6 +97,7 @@ func TestAttributesFromDSN(t *testing.T) {
 			expected: []attribute.KeyValue{
 				semconv.ServerAddress("example.com"),
 				semconv.ServerPort(3307),
+				semconv.DBNamespace("db"),
 			},
 		},
 		{
@@ -103,16 +120,25 @@ func TestAttributesFromDSN(t *testing.T) {
 			},
 		},
 		{
+			dsn: "example.com/db",
+			expected: []attribute.KeyValue{
+				semconv.ServerAddress("example.com"),
+				semconv.DBNamespace("db"),
+			},
+		},
+		{
 			dsn: "postgres://root:secret@0.0.0.0:42/db?param1=value1&paramN=valueN",
 			expected: []attribute.KeyValue{
 				semconv.ServerAddress("0.0.0.0"),
 				semconv.ServerPort(42),
+				semconv.DBNamespace("db"),
 			},
 		},
 		{
 			dsn: "postgres://root:secret@2001:db8:1234:5678:9abc:def0:0001/db?param1=value1&paramN=valueN",
 			expected: []attribute.KeyValue{
 				semconv.ServerAddress("2001:db8:1234:5678:9abc:def0:0001"),
+				semconv.DBNamespace("db"),
 			},
 		},
 		{
@@ -120,6 +146,7 @@ func TestAttributesFromDSN(t *testing.T) {
 			expected: []attribute.KeyValue{
 				semconv.ServerAddress("2001:db8:1234:5678:9abc:def0:0001"),
 				semconv.ServerPort(42),
+				semconv.DBNamespace("db"),
 			},
 		},
 		{
@@ -127,6 +154,7 @@ func TestAttributesFromDSN(t *testing.T) {
 			expected: []attribute.KeyValue{
 				semconv.ServerAddress("0.0.0.0"),
 				semconv.ServerPort(42),
+				semconv.DBNamespace("db"),
 			},
 		},
 		{
@@ -134,14 +162,30 @@ func TestAttributesFromDSN(t *testing.T) {
 			dsn: "root:secret@tcp/db?param1=value1&paramN=valueN",
 			expected: []attribute.KeyValue{
 				semconv.ServerAddress("tcp"),
+				semconv.DBNamespace("db"),
+			},
+		},
+		{
+			// DSN lacking a db-name
+			dsn: "sqlserver://user:pass@dbhost:1433",
+			expected: []attribute.KeyValue{
+				semconv.ServerAddress("dbhost"),
+				semconv.ServerPort(1433),
+			},
+		},
+		{
+			// DSN lacking a db-name, with trailing '/'
+			dsn: "postgres://user:pass@dbhost/",
+			expected: []attribute.KeyValue{
+				semconv.ServerAddress("dbhost"),
 			},
 		},
 	}
 
 	for _, tc := range testCases {
 		t.Run(tc.dsn, func(t *testing.T) {
-			got := AttributesFromDSN(tc.dsn)
-			assert.Equal(t, tc.expected, got)
+			gotAttrs := AttributesFromDSN(tc.dsn)
+			assert.Equal(t, tc.expected, gotAttrs)
 		})
 	}
 }